models/model_upload/multimodal_models/moondream2-VLM/requirements.txt
torch==2.4.0
tokenizers==0.19.1
transformers==4.44.2
torchvision==0.19.0
accelerate==0.34.2
scipy==1.10.1
optimum==1.22.0
xformers==0.0.27.post2
einops==0.8.0
pillow==10.4.0
requests==2.32.3
